requirements.txt
# Requirements para Python 3.13
# Versiones compatibles actualizadas

# Selenium - Automatización web
selenium==4.27.1
webdriver-manager==4.0.2

# OCR - Reconocimiento de texto
easyocr==1.7.2
pytesseract==0.3.13

# Procesamiento de imágenes (versiones compatibles con Python 3.13)
opencv-python==4.10.0.84
pillow==11.0.0

# Utilidades
python-dotenv==1.0.1
requests==2.32.3

# Data handling (versiones compatibles)
pandas==2.2.3
numpy==2.1.3

# Opcional pero recomendado
tqdm==4.66.5